pldm-fw: Use mctp::Listener and RespChannel
The channel from an incoming request is stored in the Responder
object so that it can be used to send outgoing requests. Users of the
Responder can retrieve the endpoint with pending_reply_ep(). Incoming
messages are provided to the Responder using request_in and reply_in
functions.

This also adds a set_state_with() helper to switch between State enum
variants.

Signed-off-by: Matt Johnston <matt@codeconstruct.com.au>