estack: Convert serial transport to packet-transfer-style API
Currently, MctpSerialHandler's public API gets passed a mctp::Stack
object, and interacts with that for packet receive/transmit.

Newer-style end-usage of the stack objects tends to use the routing
infrastructure though, which uses a simpler packet tx/rx interface to the
transport (ie, populating [u8]s rather than handling the MCTP packets
internally). This change converts the current interface to:

    pub async fn recv_async(
        &mut self,
        input: &mut impl Read,
    ) -> mctp::Result<&[u8]> {

    pub async fn send_async(
        &mut self,
        pkt: &[u8],
        output: &mut impl Write,
    ) -> mctp::Result<()> {

- which is easier to glue into the routing interface.

In doing this, convert the standalone crate to use the new
MctpSerialHandler API, where it needs to own the mctp::Stack internally.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>