mctp-estack: Remove ReceiveHandle
Cookies can now be set directly on the MctpMessage. Deferred users need
to call .retain() before dropping a a MctpMessage to avoid the
reassembler being marked as completed (that is handled inside the
Router).

Unused Reassembler slots are now stored as a State::Unused on the
Reassembler itself, rather than None in the Stack's reassembler array.
That make it more convenient for a Reassembler to set itself as unused
from the MctpMessage's drop handler.

Signed-off-by: Matt Johnston <matt@codeconstruct.com.au>