pldm-platform: Fix oem_file_classification_name
The length field is now omitted entirely when oem_classification is set
to 0. Previously in that case length was included and set to zero, which
is not the correct message format.

The name field is renamed from oem_file_name to
oem_file_classification_name, and is now an Option.

Signed-off-by: Matt Johnston <matt@codeconstruct.com.au>